All endpoints moved to routers
